Upgrade page scaffold (#704)
* added on tap

* Got rid of page scaffold unused gesture detector

* Made requrested . changes

* Updated code styling

* Updated home page

* Added safe area

* Slightly smaller textscale for smaller devices

* Edit padding

* Merge

* Added images

* finished news feed

* Working iOS

* Added images to onboarding screens

* Got rid of list of items class

List of items did essentially the same thing as PageScaffold, except it had unneccessary styling around the appbar

* Structure update

* Moved page_scaffold components into page_scaffold folder

* Added bottom bar to protect yourself

* Update about_page.dart

* fixed format

* redmade pubspec

* update cocao pods

* Update pubspec.lock

* Update pubspec.lock

* Update client/flutter/lib/components/page_scaffold/page_scaffold.dart


* Update client/flutter/lib/pages/travel_advice.dart


* Update client/flutter/lib/pages/protect_yourself.dart


* Got rid of context parameter

* made final
